requirements.txt
chromadb==0.6.2
firecrawl-py==1.9.0
gradio==5.12.0
llama-index==0.12.10
llama-index-readers-web==0.3.3
llama-index-vector-stores-chroma==0.4.1
openai==1.59.6 
python-dotenv==1.0.1
tiktoken==0.8.0